stats.py - eliminate lambda for here.
After discussing with Lewis and then later re-thinking this,
I was trying to save on creating Location objects that wouldn't be used.
Instead I was creating lambda bindings that wouldn't be used, and
then having to invoke them when needed was making the code a little bit
noisier than I would like. Python's GC is involved either way, and this
version is cleaner looking.