Enhancement: Allow NYI scenarios to have unknown tags too?
I'm not sure how I feel about this.
I'm doing it be cause I have a new NYI test based on a known bug
and if the test were implemented it would be quarantined.
But the quarantined and nyi tags are mutually exclusive.
I want to use the nyi tag because I truly have not imlemented the code yet.
However, I am concerned about having the NYI tag become a free-for-all.