Update decorators self tests (#162)
* Ensure decorators tests work as expected

* Update and reorganize decorators tests

The first change is that the names of the tests are changed to start
with the expected results. test_ok_something means the test should
result in an 'ok'. test_skipped_something means the test should
result in a 'skipped'.

The second change is that tests that should pass are decorated so
that if they happend to skip, that skip will be converted to a
failure. This way you can run the tests and be confident that as
long as there is no FAIL from a test, then things are good to go.

* Update bad decorators tests to catch error and pass the test

* Add tests for the environment_affected parameter

* Add 'wrapt' dependency to reporting

This is for testing currently, but it will be used in the source
code soon.

* Update the expected coverage JSON based on test updates

* (fix): Linter corrections

* Ignore flake8 error that shouldn't need to be ignored

* Reorder import statements to match pep8

* Parse test results to avoid false positives

* Simplify wrongful skipped test gathering logic