Enhancement: Version Bump
Now that we have transitioned from directly posting to Splunk to
leveraging the data broker, I think that this project is ready for a
wider public audience to start using. To that end, I think that it
warrants a minor version bump so that we can tell users, "just make sure
you are using at least 1.1".