QET-15: Document qecommon_tools (#219)
* QET-15: Google-Style Docs

Update all docs for qecommon_tools to conform to the Google Docs Style
for Python.

* QET-15: Use Napoleon

Napoleon is the built-in extension for Sphinx that allows documentation
to be auto-generated for Google Code Style docstrings. Also, remove the
intersphinx extension since that call was removed earlier but not
removed from the extensions list.

* QET-15: Add qecommon_tools to the PATH

In order for `autodoc` to find and document qecommon_tools, it must know
to look inside the qecommon_tools folder to find the actual library, not
the supporting files to help build and document the library.

* QET-15: Call sphinx-apidoc

Add a call to sphinx-apidoc to build the documentation for
qecommon_tools.

* QET-15: Add qecommon_tools to documentation index

Add the created qecommon_tools rst file to the documentation index.

* QET-15: Follow PEP257

Except for double quotes.

* QET-15: Alternate PEP-257 intention

Use the alternate method for PEP-257 indentation where the first line is
just the triple quote.

* QET-15: Missing letter

Add missing "h" from Otherwise.