QET-2: Manual publication of coverage date.
Shows the curl command and envelope file.
This is why I was hoping that the meta-data parameters could
be URL parameters, so that we could take a raw JSON coverage file
and upload it directly to a URL with the host name (and timestamp
if needed). Instead we have to create a wrapper. Le Sigh.

Also fix the sample JSON in coverage.rst.