QET-16 (FF): simplify logging class.
Most of the use cases for this class come from RBA,
and are based on sub-classing.  Removing the parameters for dynamic
instance method changing because that's another form of customization
that should be done with subclassing.

While this does add a bit more complexity to the test code,
the test code is more realistic because it uses sub-classes.