Enhancement: Auto generate glossaries.
As in the examples of the python builtin itertools and external
more-itertools packages, these changes allow a module writer to
tag functions and classes with arbitrary classification strings
and then create rST tables organized by those classifications.
Though not requied, the idea would be to update the module __doc__
string with this new information.